CountUpHeader 컴포넌트 ui 및 공통 코드 정리 리팩토링. 컴포넌트 내 과도한 코드 분리 및 useModal 을 사용해 중복된 모달 사용 로직 간소화
