+++ src/components/ModalPortal.tsx
@@ -3,11 +3,11 @@ import { createPortal } from 'react-dom'
 
 interface Props {
   children: ReactNode
-  closePortal: () => void
+  onClose: () => void
   isOpened: boolean
 }
 
-const ModalPortal = ({ children, closePortal, isOpened }: Props) => {
+const ModalPortal = ({ children, onClose, isOpened }: Props) => {
   useEffect(() => {
     if (isOpened) {
       document.body.style.overflow = 'hidden'
@@ -21,7 +21,7 @@ const ModalPortal = ({ children, closePortal, isOpened }: Props) => {
       <div
         className="modal-background fixed top-0 left-0 right-0 bottom-0 z-40  overflow-hidden bg-black opacity-80"
         role="presentation"
-        onClick={closePortal}
+        onClick={onClose}
       />
       <div className="fixed top-1/2 left-1/2 z-50 -translate-x-1/2 -translate-y-1/2 p-3">{children}</div>
     </div>,
