+++ src/components/timer/TimerButtons.tsx
@@ -1,0 +1,75 @@
+interface TimerButtonsProps {
+  hasStarted: boolean
+  isRunning: boolean
+  onStartClick: () => void
+  onResetClick: () => void
+  onPauseClick: () => void
+  onRestartAfterPauseClick: () => void
+}
+
+export const TimerButtons = ({
+  hasStarted,
+  isRunning,
+  onStartClick,
+  onResetClick,
+  onPauseClick,
+  onRestartAfterPauseClick,
+}: TimerButtonsProps) => {
+  if (!hasStarted)
+    return (
+      <div>
+        <button onClick={onStartClick} className="btn-primary btn h-14 gap-2 rounded-full px-5 text-lg text-white">
+          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="h-6 w-6">
+            <path
+              fillRule="evenodd"
+              d="M4.5 5.653c0-1.426 1.529-2.33 2.779-1.643l11.54 6.348c1.295.712 1.295 2.573 0 3.285L7.28 19.991c-1.25.687-2.779-.217-2.779-1.643V5.653z"
+              clipRule="evenodd"
+            />
+          </svg>
+          타이머 시작하기
+        </button>
+      </div>
+    )
+
+  return (
+    <div className="flex items-center justify-center gap-3">
+      {isRunning ? (
+        <button onClick={onPauseClick} className="btn-primary btn h-14 rounded-full px-5 text-lg text-white">
+          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="h-6 w-6">
+            <path
+              fillRule="evenodd"
+              d="M6.75 5.25a.75.75 0 01.75-.75H9a.75.75 0 01.75.75v13.5a.75.75 0 01-.75.75H7.5a.75.75 0 01-.75-.75V5.25zm7.5 0A.75.75 0 0115 4.5h1.5a.75.75 0 01.75.75v13.5a.75.75 0 01-.75.75H15a.75.75 0 01-.75-.75V5.25z"
+              clipRule="evenodd"
+            />
+          </svg>
+          멈추기
+        </button>
+      ) : (
+        <button
+          onClick={onRestartAfterPauseClick}
+          className="btn-primary btn h-14 gap-2 rounded-full px-5 text-lg text-white">
+          <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" className="h-6 w-6">
+            <path
+              fillRule="evenodd"
+              d="M4.5 5.653c0-1.426 1.529-2.33 2.779-1.643l11.54 6.348c1.295.712 1.295 2.573 0 3.285L7.28 19.991c-1.25.687-2.779-.217-2.779-1.643V5.653z"
+              clipRule="evenodd"
+            />
+          </svg>
+          재시작
+        </button>
+      )}
+
+      <button
+        onClick={onResetClick}
+        className="btn-circle btn h-14 w-14 rounded-full border border-grey-850 bg-grey-900">
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="white" className="h-8 w-8">
+          <path
+            fillRule="evenodd"
+            d="M4.5 7.5a3 3 0 013-3h9a3 3 0 013 3v9a3 3 0 01-3 3h-9a3 3 0 01-3-3v-9z"
+            clipRule="evenodd"
+          />
+        </svg>
+      </button>
+    </div>
+  )
+}
