+++ src/pages/Register.tsx
@@ -95,7 +95,7 @@ const Register = () => {
         </div>
       </div>
       {modalVisible && (
-        <ModalPortal closePortal={closeModal} isOpened={modalVisible}>
+        <ModalPortal onClose={closeModal} isOpened={modalVisible}>
           <div className="fixed right-1/2 bottom-1/2 w-[24.25rem] translate-x-1/2 translate-y-1/2 rounded-2xl bg-grey-850 px-[1.375rem] pb-[1.125rem] pt-[1.5625rem]">
             <div className="flex flex-col">
               <p className="mb-4 text-[1.375rem] font-bold leading-[140%] text-grey-200">{modalContent.modalTitle}</p>
